+++ History.md
@@ -1,7 +1,7 @@
 ### vNext ###
 - Added limited support for Rider 2017.2 (see the package description for supported features)
+- Added warnings for NUnit parameterized tests (+ quick fixes)
 - Added quick fix for "Mismatch between type and file name" to fix the file name
-- Added warnings for NUnit parameterized tests
 
 ### 0.11.0 ###
 - Added "Run Unit Tests in File" action