+++ History.md
@@ -1,4 +1,4 @@
-### vNext ###
+### 1.0.0 ###
 - Added limited support for Rider 2017.2 (see the package description for supported features)
 - Added warnings for NUnit parameterized tests (+ quick fixes)
 - Added quick fix for "Mismatch between type and file name" to fix the file name